requirements.txt
ipython==8.5.0
RPi.GPIO==0.7.1
python-dotenv==0.21.0
requests==2.28.1
